fix: navbar selection bug
